scripts/evaluate.py
# scripts/evaluate.py

import argparse
import yaml
import os
import time
import torch
from torch.utils.data import DataLoader
import logging # Import logging

# Import custom modules
from datasets.ct_denoise_dataset import CTDenoiseDataset # Assuming this is correct
# NOTE: Replace the placeholder imports below with the actual imports
# from models.archs import define_network
# from utils.logging_utils import setup_logger
# from utils.eval_utils import evaluate_model # Assuming this is correct


# Setup logger for this script
logger = logging.getLogger(__name__)
if not logger.handlers:
    # Use the utility to setup if not already configured by main execution
    from utils.logging_utils import setup_logger
    setup_logger(__name__)

# Placeholder imports - REMOVE THESE ONCE THE ACTUAL MODULES ARE READY
# These are minimal placeholders to allow the script structure to be generated.
# Replace these with the actual imports from your project structure.
logger.warning("Using placeholder imports for models and utils. Replace with actual imports.")

def define_network(opt):
    """Placeholder for defining the main network (CGNet)."""
    logger.info(f"Using placeholder define_network in evaluate.py. Defining dummy network of type: {opt['type']}")
    class DummyNetwork(nn.Module):
        def __init__(self, img_channel, width, **kwargs):
            super().__init__()
            # Simple pass-through or identity-like operation for testing
            self.identity = nn.Identity()
        def forward(self, x):
            # In a real scenario, this would be model(x)
            # For placeholder, just return the input (simulate no denoising)
            logger.debug("DummyNetwork forward pass: Returning input as output.")
            return x
    return DummyNetwork(**opt.get('args', {}))

# Need to implement evaluate_model in utils/eval_utils.py
def evaluate_model(model, dataloader, device, logger_eval, epoch=None, writer=None, metrics_list=["psnr", "ssim"]):
    """Placeholder for model evaluation function."""
    logger_eval.info("Using placeholder evaluate_model. Simulating evaluation...")
    # Simulate some dummy metric values
    avg_metrics = {"psnr": 30.0, "ssim": 0.9} # Fixed dummy values for simple test
    logger_eval.info(f"Placeholder evaluation results: {avg_metrics}")
    # Note: This placeholder does not use epoch or writer
    return avg_metrics

# END OF PLACEHOLDER


def load_model(checkpoint_path, model, device):
    """
    Loads model weights from a checkpoint.

    Args:
        checkpoint_path (str): Path to the checkpoint file.
        model (nn.Module): The model to load the state dict into.
        device (torch.device): Device to load the model onto.

    Returns:
        nn.Module: The model with loaded weights, moved to the specified device.
    """
    if not os.path.exists(checkpoint_path):
        logger.error(f"Model checkpoint not found at {checkpoint_path}. Cannot load model.")
        raise FileNotFoundError(f"Model checkpoint not found: {checkpoint_path}")

    logger.info(f"Loading model weights from {checkpoint_path}")
    try:
        # Load checkpoint to CPU first, then move to device
        checkpoint = torch.load(checkpoint_path, map_location='cpu')
        # Load model state dict, allowing for potential strict=False if model architecture changed slightly
        model.load_state_dict(checkpoint['model_state_dict'], strict=True) # Use strict=True by default
        logger.info("Model weights loaded successfully.")
    except Exception as e:
        logger.error(f"Error loading model weights from {checkpoint_path}: {e}")
        logger.error("Evaluation cannot proceed without loading model weights.")
        raise e # Re-raise the exception

    model.to(device) # Move model to the target device
    return model


def main(config_path, checkpoint_path):
    """Main evaluation function for the CGNet model."""
    with open(config_path, 'r') as f:
        config = yaml.safe_load(f)

    # Use the relevant sections of the config
    model_config = config['model']
    dataset_config = config['dataset']
    evaluation_config = config['evaluation']
    # Evaluation script doesn't need loss, optimizer, or training configs


    # --- Setup ---
    # Setup logging - use the utility function
    # Log to console by default for evaluation script
    logger = setup_logger('evaluate_logger', stream=True, log_file=None, level=logging.INFO)
    logger.info(f"Starting CGNet evaluation using checkpoint: {checkpoint_path}")
    logger.info(f"Config:\n{yaml.dump(config, indent=4)}") # Log the full config

    # Device setup
    # Use device from config or default to cuda if available, else cpu
    # Use the device specified in the training config section if available, otherwise default
    device = torch.device(config.get('training', {}).get('device', 'cuda') if torch.cuda.is_available() else 'cpu')
    logger.info(f"Using device: {device}")


    # --- Data Loading ---
    # Load test dataset and dataloader
    # We need both LDCT (input) and NDCT (target) for evaluation
    test_dataset = CTDenoiseDataset(
        root=dataset_config['args']['root'],
        mode='test', # Evaluate on the test set
        transform=None # Use default ToTensor transform
    )

    if len(test_dataset) == 0:
        logger.error("No data found in the test set for evaluation. Please check the dataset path and preprocessing.")
        return

    test_dataloader = DataLoader(
        test_dataset,
        batch_size=dataset_config.get('test_batch_size', 1), # Typically batch size 1 for evaluation
        shuffle=False, # No need to shuffle test data
        num_workers=dataset_config.get('num_workers', 4),
        pin_memory=True and device.type == 'cuda'
    )
    logger.info(f"Test DataLoader created with {len(test_dataloader)} batches.")


    # --- Model Setup ---
    # Define the main network (CGNet)
    # NOTE: Ensure models.archs.__init__.py and the CGNet architecture file (CGNet_arch.py) are ready
    try:
         from models.archs import define_network as actual_define_network
         model = actual_define_network(model_config) # Define model (weights not loaded yet)
         logger.info("Successfully imported and defined CGNet model structure.")
    except ImportError:
         logger.error("Could not import actual define_network. Using placeholder.")
         model = define_network(model_config)
    except Exception as e:
         logger.error(f"Error defining CGNet model structure: {e}. Using placeholder.")
         model = define_network(model_config)


    # Load the trained weights from the checkpoint and move model to device
    try:
        model = load_model(checkpoint_path, model, device)
    except FileNotFoundError:
        logger.error("Exiting due to missing model checkpoint.")
        return
    except Exception as e:
        logger.error(f"Exiting due to error loading model: {e}")
        return


    # --- Evaluation ---
    # Get the list of metrics to calculate from the evaluation config
    metrics_list = evaluation_config.get('metrics', ["psnr", "ssim"])
    if not metrics_list:
        logger.warning("No metrics specified in the evaluation config. No metrics will be calculated.")

    # Use the evaluate_model utility function to perform evaluation
    # NOTE: Ensure utils.eval_utils.evaluate_model is ready
    try:
        from utils.eval_utils import evaluate_model as actual_evaluate_model
        avg_metrics = actual_evaluate_model(
            model=model,
            dataloader=test_dataloader,
            device=device,
            logger_eval=logger, # Pass the logger instance
            metrics_list=metrics_list
            # epoch and writer are not needed for standalone evaluation script
        )
        logger.info("Evaluation completed successfully using actual evaluate_model.")
    except ImportError:
         logger.error("Could not import actual evaluate_model. Using placeholder.")
         avg_metrics = evaluate_model(
             model=model,
             dataloader=test_dataloader,
             device=device,
             logger_eval=logger,
             metrics_list=metrics_list
         )
    except Exception as e:
         logger.error(f"Error during evaluation using evaluate_model: {e}. Using placeholder.")
         avg_metrics = evaluate_model(
             model=model,
             dataloader=test_dataloader,
             device=device,
             logger_eval=logger,
             metrics_list=metrics_list
         )


    # Report final average metrics
    logger.info("\n--- Final Average Evaluation Metrics ---")
    if avg_metrics:
        for metric, avg_val in avg_metrics.items():
            logger.info(f"  Average {metric.upper()}: {avg_val:.4f}")
    else:
        logger.info("No metrics were calculated.")

    logger.info("CGNet evaluation finished.")


if __name__ == "__main__":
    parser = argparse.ArgumentParser(description='Evaluate the trained CGNet denoiser.')
    parser.add_argument('--config', type=str, default='configs/default_config.yaml',
                        help='Path to the configuration file (YAML).')
    parser.add_argument('--checkpoint', type=str, required=True,
                        help='Path to the trained CGNet model checkpoint (.pth file).')
    # Output directory is not strictly needed for evaluation metrics, but could be added
    # if you wanted to save evaluated images or a separate log file here.
    # parser.add_argument('--output_dir', type=str, default=None,
    #                     help='Optional directory to save evaluation results (e.g., log file).')

    args = parser.parse_args()

    main(args.config, args.checkpoint) # Pass output_dir if added to args

