requirements.txt
torch>=1.8.0
torchvision>=0.9.0
PyYAML>=5.4.1
numpy>=1.20.0
Pillow>=8.0.0
pydicom>=2.2.0
pytorch-msssim>=0.1
tensorboard>=2.5.0
scikit-image 
# Add other dependencies as needed (e.g., for specific AE architecture)